+++ src/app/follow/page.tsx
@@ -1,29 +1,9 @@
 "use client";
-import { Box, Tabs, Tab, createTheme, ThemeProvider } from "@mui/material";
+import { createTheme } from "@mui/material";
 import { getUuidFromCookie } from "@/actions/users";
 import { useQuery } from "@tanstack/react-query";
 import { useSearchParams } from "next/navigation";
-import FollowUsers from "@/components/like/FollowUsers";
-import FollowerUsers from "@/components/like/FollowerUsers";
-import MutualUsers from "@/components/like/MutualUsers";
-
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: "#22d3ee",
-    },
-  },
-  components: {
-    MuiTab: {
-      styleOverrides: {
-        root: {
-          fontSize: "1.2rem", // タブの文字サイズを大きくする
-          padding: "1rem", // タブのパディングを増やす
-        },
-      },
-    },
-  },
-});
+import FollowPage from "@/components/follow/FollowPage";
 
 const Page = () => {
   const value = useSearchParams().get("tab") || "follows";
@@ -52,36 +32,7 @@ const Page = () => {
     },
   });
 
-  const handleChange = (event: React.SyntheticEvent, newValue: string) => {
-    const searchParams = new URLSearchParams(window.location.search);
-    searchParams.set("tab", newValue);
-    const newUrl = `${window.location.pathname}?${searchParams.toString()}`;
-    window.history.pushState(null, "", newUrl);
-  };
-
-  return (
-    <ThemeProvider theme={theme}>
-      <Box sx={{ width: "100%", bgcolor: "background.paper" }}>
-        <Tabs
-          value={value}
-          onChange={handleChange}
-          centered
-          sx={{ color: "#22d3ee" }}
-        >
-          <Tab label="フォロー中" value="follows" />
-          <Tab label="相互フォロー" value="mutual" />
-          <Tab label="フォロワー" value="followers" />
-        </Tabs>
-      </Box>
-      {data && value === "follows" && <FollowUsers follows={data.follows} />}
-      {data && value === "mutual" && (
-        <MutualUsers follows={data.follows} followers={data.followers} />
-      )}
-      {data && value === "followers" && (
-        <FollowerUsers followers={data.followers} />
-      )}
-    </ThemeProvider>
-  );
+  return <FollowPage data={data} isMe={true} />;
 };
 
 export default Page;
